2nd ontology/requirements.txt
flask==3.0.0
pandas>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
rdflib>=6.0.0
numpy>=1.20.0
jsonschema>=4.17.0
